[fixup] Change date presentation for untagged commits
For the 'public' branch of this repository, I don't really need to use
tags. I can just expose the date and the commit SHA and that should be
fine. Tags can be reserved for submissions or other versions that *I*
want to release and keep track of on a private branch.