[#18] Relax assumption on dispersion of mu_t in Theorem 1
This commit makes a fairly large change to the results. The assumption on
the mu_t's turns out to be very restrictive, so I've relaxed it. To do this,
I've added a separate lemma that shows that the dispersion component of X*-bar
converges to zero at the right rate to not affect the asymptotic distribution.
Then I've added a small part at the beginning of Theorem 1's proof, which
basically points to Lemma 3, and modified the proof of (old numbering) Lemma 5,
which proves uniform integrability of the bootstrapped blocks.