[#18] Strengthen one component of Lemma 1
The last part of Lemma 1 deals with convergence of \sum M_j^2 / n^2. To
relax the strong assumption on near homogeneity of the means that I'm
currently making, I need a slightly different version of this convergence;
I want the rate that \sum M_j^2 diverges. This commit changes the result
to give this stronger conclusion.