[#27] Add name of last version of paper to ignore list
It's useful to have a version of the last submitted version of the
paper to generate the LaTeX diff file. But we don't want to add it
to the repository, so I'm adding it to the repo's ignore list.

(cherry picked from commit 17798a42b26bcb0c415ef2316075e32b86151f00)